+++ Bridge/PortalNode/Configuration/Contract/InstructionLoaderInterface.php
@@ -4,12 +4,12 @@ declare(strict_types=1);
 
 namespace Heptacom\HeptaConnect\Core\Bridge\PortalNode\Configuration\Contract;
 
+use Heptacom\HeptaConnect\Core\Bridge\PortalNode\Configuration\InstructionTokenCollection;
+
 interface InstructionLoaderInterface
 {
     /**
      * Load instruction tokes to alter portal node configurations.
-     *
-     * @return InstructionTokenContract[]
      */
-    public function loadInstructions(): array;
+    public function loadInstructions(): InstructionTokenCollection;
 }