+++ CHANGELOG.md
@@ -98,6 +98,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Make class final: `\Heptacom\HeptaConnect\Core\Portal\ComposerPortalLoader`
 - Change return type of `\Heptacom\HeptaConnect\Core\Reception\ReceiveContext::getEventDispatcher` from `\Psr\EventDispatcher\EventDispatcherInterface` to `\Symfony\Component\EventDispatcher\EventDispatcherInterface` in order to allow registering listeners
 - Change parameter of type `\League\Flysystem\FilesystemInterface` in `\Heptacom\HeptaConnect\Core\Storage\Normalizer\StreamNormalizer::__construct` and `\Heptacom\HeptaConnect\Core\Storage\Normalizer\StreamDenormalizer::__construct` to string representing a directory
+- Change return type of `\Heptacom\HeptaConnect\Core\Bridge\PortalNode\Configuration\Config::buildInstructions`, `\Heptacom\HeptaConnect\Core\Bridge\PortalNode\Configuration\Contract\InstructionLoaderInterface::loadInstructions` and `\Heptacom\HeptaConnect\Core\Bridge\PortalNode\Configuration\InstructionFileLoader::loadInstructions` to new type safe collection `\Heptacom\HeptaConnect\Core\Bridge\PortalNode\Configuration\InstructionTokenCollection`
 
 ### Deprecated